Adhere to development standards
