+++ torcheeg/models/__init__.py
@@ -6,4 +6,3 @@ from .gnn import *
 from .rnn import *
 from .transformer import *
 from .vae import *
-from .any import *
