+++ torcheeg/models/cnn/__init__.py
@@ -1,7 +1,8 @@
 from .ccnn import CCNN
 from .eegnet import EEGNet
 from .fbccnn import FBCCNN
-from .fbcnet import FBCNet,FBMSNet
+from .fbcnet import FBCNet
+from .fbmsnet import FBMSNet,Conv2dWithConstraint,MixedConv2d,LinearWithConstraint,SamePadConv2d
 from .mtcnn import MTCNN
 from .stnet import STNet
 from .tsception import TSCeption
