+++ torcheeg/trainers/__init__.py
@@ -1,4 +1,5 @@
-from .classifier import ClassifierTrainer,CLossClassifierTrainer
+from .classifier import ClassifierTrainer
+from .closs_classifier import CLossClassifierTrainer
 from .domain_adaption import *
 from .generative import *
 from .self_supervised import *
